+++ psl-core/src/main/java/org/linqs/psl/application/learning/weight/gradient/onlineminimizer/OnlineSquaredError.java
@@ -51,7 +51,7 @@ public class OnlineSquaredError extends OnlineMinimizer {
                 continue;
             }
 
-            supervisedLoss += Math.pow(proxRuleObservedAtoms[rvAtomIndexToProxRuleIndex.get(atomIndex)].getValue() - observedAtom.getValue(), 2.0f);
+            supervisedLoss += (float) Math.pow(proxRuleObservedAtoms[rvAtomIndexToProxRuleIndex.get(atomIndex)].getValue() - observedAtom.getValue(), 2.0f);
         }
 
         return supervisedLoss;
